+++ Core/Maps/Geometry/SectorFlat.cs
@@ -12,13 +12,13 @@ namespace Helion.Maps.Geometry
         public double Z;
         public double PrevZ;
         public CIString Texture;
-        public byte LightLevel;
+        public short LightLevel;
         private Sector? m_sector;
 
         public Sector Sector => m_sector ?? throw new NullReferenceException("Forgot to set sector for sector flat");
         public float UnitLightLevel => LightLevel / 255.0f;
 
-        public SectorFlat(int id, CIString texture, double z, byte lightLevel, SectorFlatFace facing)
+        public SectorFlat(int id, CIString texture, double z, short lightLevel, SectorFlatFace facing)
         {
             Id = id;
             Facing = facing;
